Test tree: the two E0502 borrows and the u128 timestamp, located by the surface that now carries them
Run 51's surface paired each error with its `--> file:line:col`, which is the
difference between three annotations naming symptoms and three annotations naming
addresses. Clearing the resolution errors (missing `use` for `Executor`, the
`let _ = challenge_id;` silencer) let rustc reach borrowck/typeck and expose what
had been suppressed:

- src/chain/storage_economics_tests.rs 525/529, E0502 x2: `all_reallocation_tickets`
  returns `Vec<&Ticket>`, and the test held that reference across
  `state.add_balance` and `accept_storage_reallocation_with_escrow`. The field the
  test reads later (`ticket_id`) is now copied out of a scoped block; the whole ticket
  is not cloned, because the clone would keep the borrow alive to no purpose and hide
  which field the assertions depend on.
- src/tests/pow_light_client.rs 222, E0308: `timestamp_ms` is `u128`, and
  `1_000 + height` infers `u64` from `(1u64..=4)`. Now `1_000 + u128::from(height)`.

Also recorded: the Typos check on this PR is not this branch's doing. PR CI runs on
the merge ref, and `main` (d07225d) carries
`01a08674-...-b135-24e6cc8e1539 (2).patch` at the repo root - 2.3 MB, 37k lines,
added by "Add files via upload", referenced by nothing. typos 1.48.0 on a worktree
of origin/main reports exactly the 6 findings run 51 showed once my patch mirrors
were excluded, all six inside that file. The fix is `git rm` on main, not a
whitelist entry here. (Sandbox fact for the next reader:
release-assets.githubusercontent.com is unreachable from here, like the Actions log
host, so the pinned binary cannot be downloaded - the same version from PyPI can.)

Not compiled or type-checked locally: no toolchain in this sandbox. Whether the
lib-test target builds is run 52's answer, and run 52 is also the first run where
the gate suite runs outside the reach of the Format step.

+++ src/tests/pow_light_client.rs
@@ -219,7 +219,9 @@ fn pow_fixture(first_state_root: [u8; 32]) -> (
                 state_root,
                 tx_root: [2u8; 32],
                 event_root: [5u8; 32],
-                timestamp_ms: 1_000 + height,
+                // `height` is the `u64` from the range, and this field is `u128`:
+                // `1_000 + height` infers `u64` and the field then refuses it (E0308).
+                timestamp_ms: 1_000 + u128::from(height),
                 nonce: 0,
                 difficulty_bits: 4,
             },
